backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-cors-headers==4.4.0
paho-mqtt==2.1.0
drf-spectacular==0.27.2
daphne==4.1.2
whitenoise==6.7.0

